Moves transmission effect into transmission module
